+++ mne/io/edf/edf.py
@@ -1086,7 +1086,7 @@ def read_raw_edf(input_fname, montage=None, eog=None, misc=None,
         None is accepted as an alias for False.
 
         .. warning:: 0.18 does not allow for stim channel synthesis from
-                     the TAL channel called 'EDF Annotations' or
+                     the TAL channelsnel called 'EDF Annotations' or
                      'BDF Annotations'. The TAL channel is parsed
                      and put in the raw.annotations attribute.
                      Use :func:`mne.events_from_annotations` to obtain
@@ -1128,8 +1128,8 @@ def read_raw_edf(input_fname, montage=None, eog=None, misc=None,
     Use :func:`mne.events_from_annotations` to obtain events from the
     annotations.
 
-    If channels named 'Status' or 'STATUS' will be considered analog stim
-    channels, use method ``raw.find_edf_events`` to recover the events encoded
+    Channels that are named 'Status' or 'STATUS' will be considered analog stim
+    channels. Use method ``raw.find_edf_events`` to recover the events encoded
     in this channel.
 
     See Also